adding interfaces to repo, service, controller
