+++ src/main/java/pl/wiktor/management/user/service/ActiveUserService.java
@@ -10,7 +10,7 @@ import pl.wiktor.management.user.model.enums.TableSearcher;
 import pl.wiktor.management.user.exception.AccountLoginException;
 import pl.wiktor.management.user.exception.AccountPasswordException;
 import pl.wiktor.management.user.model.mapper.Mapper;
-import pl.wiktor.management.user.repository.AccountRepository;
+import pl.wiktor.management.user.repository.AccountRepositoryImpl;
 import pl.wiktor.management.user.repository.ActiveUserRepository;
 
 import java.util.UUID;
@@ -18,11 +18,11 @@ import java.util.UUID;
 @Service
 public class ActiveUserService {
     private ActiveUserRepository activeUserRepository;
-    private AccountRepository accountRepository;
+    private AccountRepositoryImpl accountRepository;
     private final Logger logger = LoggerFactory.getLogger(this.getClass());
 
     @Autowired
-    public ActiveUserService(ActiveUserRepository activeUserRepository, AccountRepository accountRepository) {
+    public ActiveUserService(ActiveUserRepository activeUserRepository, AccountRepositoryImpl accountRepository) {
         this.activeUserRepository = activeUserRepository;
         this.accountRepository = accountRepository;
     }
